requirements.txt
# Django framework
Django==5.2.7

# REST framework
djangorestframework==3.15.2

# CORS headers for frontend-backend communication
django-cors-headers==4.6.0

# Environment variable management (optional if you want to keep dotenv support)
python-dotenv==1.0.1

# Optional: Gunicorn for production deployment
gunicorn==21.2.0

# Optional: Pillow for image handling in UserProfile/Game
Pillow==10.1.0
